chore:renamed App component to Playground
