+++ playground/src/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './App';
+import Playground from './Playground';
 
 import '@fortawesome/fontawesome-svg-core/styles.css';
 import '@kyma-project/asyncapi-react/lib/styles/fiori.css';
 import './common/icons';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(<Playground />, document.getElementById('root'));
